Create the first Ansible Check to test new macros
It was chosen the display_login_attempts rule for the test because this
rule is relatively complex in terms of checks. It is necessary to assert
a file and its content using regex. It is also important to ensure no
duplication is present because this can change the PAM behavior
unexpectedly.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>