Introduce Ansible Macros for Ansible Checks
The goal is to give high scalability and speed for creation of Ansible
Checks without impact on quality and safety. These macros are exclusive
for Ansible Checks and the number of Ansible modules on them is very
limited because we can basically check everying with very few modules
plus the ansible facts. Therefore, the modules used here don't modify
the system. Instead, they are only capable to collect and assess
information. The macros are as granular as possible and share the same
standards, so we can very easily create new macros on demand basically
with a combination of some blocks.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>