Create code to compile Ansible Checks
This initial version of build_checks.py was created to put everything
together. The first version was created for Ansible Checks but the code
was designed with scalability and flexibility in mind. The intention is
to make it easy to adopt new checks in the future. There are some
important new concepts introduced with Ansible Checks. First, they are
completely independent of SCAP, OVAL and XCCDF. Ansible Checks use
native Ansible capabilities. It means that all conditions to variables
and products properties are present in the Playbook and are assessed in
run-time. This concept opens a lot of new possibilities, such as
checking multiple different systems at the same time, much easier
interpretation of assessment, much more performance and resource
optimizations.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>